third_party/blink/web_tests/virtual/threaded/http/tests/worklet/paint-worklet-csp-eval-expected.txt

CONSOLE ERROR:  Hash of blocked script: "eval-sha256-wnpR9Zq+PnzDQHy06uKSjdAKPXRJtnJIZlppUZwnazI=".
CONSOLE ERROR: Uncaught EvalError: Refused to evaluate a string as JavaScript because 'unsafe-eval' is not an allowed source of script in the following Content Security Policy directive: "script-src 'self' 'unsafe-inline'".

CONSOLE ERROR:  Hash of blocked script: "eval-sha256-wnpR9Zq+PnzDQHy06uKSjdAKPXRJtnJIZlppUZwnazI=".
CONSOLE ERROR: Uncaught EvalError: Refused to evaluate a string as JavaScript because 'unsafe-eval' is not an allowed source of script in the following Content Security Policy directive: "script-src 'self' 'unsafe-inline'".

CONSOLE ERROR:  Hash of blocked script: "eval-sha256-wnpR9Zq+PnzDQHy06uKSjdAKPXRJtnJIZlppUZwnazI=".
CONSOLE ERROR: Uncaught EvalError: Refused to evaluate a string as JavaScript because 'unsafe-eval' is not an allowed source of script in the following Content Security Policy directive: "script-src 'self' 'unsafe-inline'".

CONSOLE ERROR:  Hash of blocked script: "eval-sha256-wnpR9Zq+PnzDQHy06uKSjdAKPXRJtnJIZlppUZwnazI=".
CONSOLE ERROR: Uncaught EvalError: Refused to evaluate a string as JavaScript because 'unsafe-eval' is not an allowed source of script in the following Content Security Policy directive: "script-src 'self' 'unsafe-inline'".

This is a testharness.js-based test.
Harness: the test ran to completion.
